refactor: changed some remaining pandas to polars & did cast64 and remove some codes

+++ Appearance_rate_V2_batch.py
@@ -433,7 +433,7 @@ class HonkaiStatistics_V2_Batch:
         per_eidolon = (
             get_performance_stats(base_data, ["Character", "cons"])
             # OPTIMIZATION: pl.format is vastly faster than cast + concat
-            .with_columns(pl.format("Eidolon {}", pl.col("cons")).alias("Eidolon_Level"))
+            .with_columns(pl.format("Eidolon {}", pl.col("cons").cast(pl.Int64)).alias("Eidolon_Level"))
             
         )
