+++ Appearance_rate_V2_batch_anomaly.py
@@ -423,7 +423,7 @@ class HonkaiStatistics_V2_Anomaly_Batch:
         per_eidolon = (
             get_performance_stats(base_data, ["Character", "cons"])
             # OPTIMIZATION: pl.format is vastly faster than cast + concat
-            .with_columns(pl.format("Eidolon {}", pl.col("cons")).alias("Eidolon_Level"))
+            .with_columns(pl.format("Eidolon {}", pl.col("cons").cast(pl.Int64)).alias("Eidolon_Level"))
             
         )
